Rename utility file to avoid core package conflicts
