+++ tests/test_uuid.py
@@ -1,7 +1,7 @@
 import uuid
 
 from le_utils.constants.uuid import KOLIBRI_ECOSYSTEM_UUID_NAMESPACE
-from le_utils.uuid import generate_ecosystem_namespaced_uuid
+from le_utils.uuidv5 import generate_ecosystem_namespaced_uuid
 
 
 def test_namespace_does_not_change():
